[DF] Different head nodes for different data sources
In distributed RDataFrame, the head node of the RDF computation graph also stores information about the user provided arguments to the RDF constructor, i.e. information about the source of data.
The current implementation has only one HeadNode class responsible for distinguishing between the two supported data sources, i.e. empty RDF with sequential entries and TTree/TChain based RDF.
This commit introduces instead multiple head node types, one per data source. In particular, there is now a `EmptySourceHeadNode` and a `TreeHeadNode`.
To decide which is the correct type of head node for a particular distributed dataframe, this commit introduces a factory function `get_headnode` that parses the user provided arguments to the RDataFrame constructor and returns the correct head node instance.

It would follow that each data source might have different information to send to the distributed resources. This is the case, and it will be addressed in the future with the introduction of a different Range object per each head node type

These change brought a few more changes along:
* In the `get_headnode` factory function there is an early check for the passed arguments to the RDF constructor. Improve the check by catching the `TypeError` that comes from the C++ overload failing to instantiate an RDF with wrong constructor arguments, by raising a new `TypeError` with a more user friendly message.
* The logic of the distributed execution present in the Base backend and in the Spark backend need to be changed in order to accomodate the different head node types.
* The presence of different head node types collides with the current implementation of the `RangesBuilder` class that expected a single type of headnode to retrieve attributes from. Workaround this by using `try except` clauses in the properties of `RangesBuilder`. It is an ugly workaround that will be superseded very soon when also the creation of the ranges will be separated according to the data source.
